moved card files to (templates) folder
