+++ src/app/(templates)/card.module.css
@@ -7,7 +7,7 @@
 }
 
 .text {
-  padding-block: 20px 15px;
+  padding-block: 10px 15px;
   display: flex;
   align-items: baseline;
   justify-content: space-between;
@@ -33,6 +33,7 @@
   font-weight: 400;
   font-style: italic;
   font-size: 11pt;
+  line-height: 1.3rem;
   color: rgb(119, 119, 119);
   text-align: right;
   padding-left: 20px;
