+++ src/app/(templates)/card.tsx
@@ -2,17 +2,23 @@
 import styles from './card.module.css'
 import Link from 'next/link'
 import Image from 'next/image'
+import articles from '../../JournalEntries.json'
+import pic from 'public/test.jpg'
 
 
-export default function Card({title = "title", genre = "GENRE", image = "", preview = "" }) { //take in an image id
+// export default function Card({title = "title", genre = "GENRE", image=0}) { //take in an image id
+export default function Card({id=-1}) {
+    const title = id === -1 ? "Article Title" : articles[id].Title
+    const genre = id === -1 ? "GENRE" : articles[id]['Content Type'].toUpperCase()
     return (
-        <Link href="/" className={styles.card}>
-            {/* if there is no image id, then display a default color block */}
-            { (image === "") && <div className={styles.cardColor} />} 
-            { (image !== "") &&
+        <div className={styles.card}>
+            {/* if there is no article id, then display a default color block */}
+            { (id === -1) && <div className={styles.cardColor} />} 
+            { (id != -1) &&
                 <Image 
                     className={styles.cardImage}
-                    src={image}
+                    // src={articles[id].Image as any}
+                    src={pic}
                     width={400}
                     height={400}
                     alt='article image'/>
@@ -21,7 +27,5 @@ export default function Card({title = "title", genre = "GENRE", image = "", prev
                 <h2>{title}</h2>
                 <h3>{genre}</h3>
             </div>
-            { (preview !== "") && <p>{preview}</p> }
-            <p className={styles.readmore} >Read more</p>
-        </Link>
+        </div>
 )}
