--- src/app/card.module.css
@@ -1,78 +1,0 @@
-.card {
-  width: 20vw;
-  max-width: 22vw;
-  height: 24vw;
-  flex-grow: 1;
-  margin-block: 2rem;
-}
-
-.text {
-  padding-block: 20px 15px;
-  display: flex;
-  align-items: baseline;
-  justify-content: space-between;
-}
-
-/* article title */
-.card h2 {
-  font-family: var(--font-ebgaramond);
-  font-size: 16pt;
-  font-weight: 600;
-  text-align: left;
-
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-/* genre label */
-.card h3 {
-  font-family: 'Arial';
-  font-weight: 400;
-  font-style: italic;
-  font-size: 11pt;
-  color: rgb(119, 119, 119);
-  text-align: right;
-  padding-left: 20px;
-}
-
-/* preview text */
-.card p {
-  font-size: 14pt;
-  height: 70px;
-  margin-bottom: 1rem;
-
-  display: -webkit-box;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-.readmore {
-  opacity: .5;
-}
-
-.card:hover .readmore {
-  opacity: 1;
-  transition: .3s;
-  letter-spacing: .2px;
-}
-
-/* default color card */
-.cardColor {
-  background-color: var(--agora-green);
-  width: 100%;
-  height: 60%;
-}
-
-/* article image */
-.cardImage {
-  width: 100%;
-  height: 60%;
-  object-fit: cover;
-  overflow: hidden;
-}
-
